started making guide pathfind
